[REFACTOR] refactor normalization persistance providers
